python/requirements.txt
requests>=2.32.0
jaydebeapi>=1.2.3
JPype1>=1.5.0
fastapi>=0.110.0
uvicorn>=0.23.0
zeep>=4.2.1
python-dotenv>=1.0.1
